+++ src/components/Landing/Sections/Faqs.jsx
@@ -82,8 +82,7 @@ class FAQs extends Component {
         return (
             <div style={{ backgroundColor: theme.secondary[1], color: "white", boxShadow: "0px 0px 10px rgba(0, 0, 0, 0.5)", padding: 50, marginTop: 25 }}>
                 <div style={{ position: "absolute", left: "calc(15px)", top: 0, height: "calc(100%)", backgroundColor: theme.accent[1], width: 10 }}></div>    
-                <h1 style={{ marginBottom: 100 }} className="display-4 theme-font">FAQs</h1>
-                <div style={{ position: "absolute", left: "calc(15px)", top: 150, width: "calc(100% - 30px)", backgroundColor: theme.accent[1], height: 10 }}></div>    
+                <h1 style={{ marginBottom: 100 }} className="display-4 theme-font">FAQs</h1>  
                 <Container fluid>
                     <Row>
                         <Col xs={cols}>
